feat: 좋아요 기능 구현

+++ GoormRunner-BackEnd-feat-market/src/main/java/goorm_runner/backend/market/presentation/MarketController.java
@@ -124,7 +124,7 @@ public class MarketController {
                 market.getTitle(),
                 market.getContent(),
                 market.getPrice(),
-                market.getLikeCount().intValue(),
+                market.getLikeCount(),
                 market.getStatus(),
                 market.getCreatedAt(),
                 market.getUpdatedAt()
